+++ packages/adhere/changelog/CHANGELOG.md
@@ -8,6 +8,8 @@
 * ssr的支持
 * **adhere-util-dict**
   - 如果value是函数加入是否缓存的设置
+* **adhere-util-decorators**
+  - componentDidCatch中加入Loading chunk重新加载
 
 # 1.0.12